Update db_init_v2.sql
error fix

+++ api/db_init_v2.sql
@@ -1,6 +1,6 @@
 CREATE TABLE all_in_one (
     ID INT AUTO_INCREMENT PRIMARY KEY ,
     time_stamp VARCHAR(65535),
-    magnitude FLOAT(255),
+    magnitude FLOAT(23),
     City VARCHAR(65535)
 );
